[tosoApp]createTodoを分解しtodoListsの書き方変更
