+++ 03_todoApp/src/css/style.css
@@ -209,3 +209,7 @@ body {
 .finished.hide {
   display: none;
 }
+
+.displayNone {
+  display: none;
+}
