+++ 03_todoApp/src/javascript/modules/checkOverdue.js
@@ -1,0 +1,6 @@
+export const checkOverdue = (due) => {
+  const today = new Date(Date.now());
+  const deadline = new Date(due);
+  const overdue = today > deadline;
+  return overdue;
+}
