--- 03_todoApp/src/javascript/modules/displayNone.js
@@ -1,3 +1,0 @@
-export const displayNone = (item) => {
-  item.style.display = 'none';
-}
