+++ 03_todoApp/src/javascript/modules/loadListsArray.js
@@ -1,0 +1,6 @@
+export const loadListsArray = () => {
+  const array = localStorage.hasOwnProperty('todoLists')
+    ? JSON.parse(localStorage.getItem('todoLists'))
+    : []
+  return array;
+}
