+++ 03_todoApp/src/javascript/modules/resetButton.js
@@ -1,12 +1,14 @@
-import { todoLists } from "./renderLists.js"
-
-const resetButton = document.querySelector('[data-js="resetButton"]');
+import { loadListsArray } from "./loadListsArray.js";
 
 export const resetButtonClickListener = () => {
+  const resetButton = document.querySelector('[data-js="resetButton"]');
+
   resetButton.addEventListener('click', () => {
     const result = confirm('本当に削除しますか？');
     if (result) {
       localStorage.clear();
+
+      const todoLists = loadListsArray();
       todoLists.length = 0;
 
       const box = document.querySelector('[data-js="todoListBox"]');
